Allow retrieved neighbors to be filtered by category/tag id
This adds a `tag` field to each Node in the trees, which allows items
to be filtered by their tag id during retrieval. For the moment, this
still amounts post-filtering the neighbors, but the filtering happens
inside the tree traversal during retrieval.

The plan moving forward is to replace the single numeric tag field with
a list (or bitset) of tags, which has two benefits:
- it allows item membership in multiple categories or tags
- it allows parent nodes to track which categories or tags their
children belong to (in aggregate)

Setting up a situation where parents know which categories/tags their
children belong to allows intelligent tree traversal which skips over
nodes where none of the children belong to the desired category/tag,
which makes the `search_k` budget go farther in terms of finding
relevant neighbors.

Using bitsets to track category/tag membership also has the advantage of
allowing fast comparisons, which should make it possible to make
fast multi-tag queries (in either of AND/OR modes.)

The current state of this commit breaks the binary compatibility with
past indices, so compatibility tests are disabled. Seems like the ideal
situation would be to have index loading methods that can detect which
index format they're reading and work seamlessly with tagged/tagless
index files. Wanted to get some basic functionality working before
trying to tackle that, though. May just end up with separate save/load
methods for each index format.